chore: move profiles and tiers into profile module
