+++ cli/cmd/resources/profiles.go
@@ -4,26 +4,13 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/odigos-io/odigos/cli/cmd/resources/profiles"
+	cliprofiles "github.com/odigos-io/odigos/cli/cmd/resources/profiles"
 	"github.com/odigos-io/odigos/cli/cmd/resources/resourcemanager"
 	"github.com/odigos-io/odigos/cli/pkg/kube"
 	"github.com/odigos-io/odigos/common"
-	k8sprofiles "github.com/odigos-io/odigos/k8sutils/pkg/profiles"
-	commonprofiles "github.com/odigos-io/odigos/profiles"
+	"github.com/odigos-io/odigos/profiles"
+	"github.com/odigos-io/odigos/profiles/profile"
 )
-
-func GetAvailableCommunityProfiles() []commonprofiles.Profile {
-	return []commonprofiles.Profile{k8sprofiles.SemconvUpgraderProfile, k8sprofiles.CopyScopeProfile, k8sprofiles.DisableNameProcessorProfile,
-		k8sprofiles.SizeSProfile, k8sprofiles.SizeMProfile,
-		k8sprofiles.SizeLProfile, k8sprofiles.AllowConcurrentAgents}
-}
-
-func GetAvailableOnPremProfiles() []commonprofiles.Profile {
-	return append([]commonprofiles.Profile{k8sprofiles.FullPayloadCollectionProfile, k8sprofiles.DbPayloadCollectionProfile, k8sprofiles.CategoryAttributesProfile,
-		k8sprofiles.HostnameAsPodNameProfile, k8sprofiles.JavaNativeInstrumentationsProfile, k8sprofiles.JavaEbpfInstrumentationsProfile, k8sprofiles.KratosProfile, k8sprofiles.QueryOperationDetector,
-		k8sprofiles.SmallBatchesProfile},
-		GetAvailableCommunityProfiles()...)
-}
 
 func GetResourcesForProfileName(profileName common.ProfileName, tier common.OdigosTier) ([]kube.Object, error) {
 	allAvailableProfiles := GetAvailableProfilesForTier(tier)
@@ -31,7 +18,7 @@ func GetResourcesForProfileName(profileName common.ProfileName, tier common.Odig
 		if p.ProfileName == common.ProfileName(profileName) {
 			if p.KubeObject != nil {
 				filename := fmt.Sprintf("%s.yaml", profileName)
-				return profiles.GetEmbeddedYAMLFileAsObjects(filename, p.KubeObject)
+				return cliprofiles.GetEmbeddedYAMLFileAsObjects(filename, p.KubeObject)
 			}
 			if len(p.Dependencies) > 0 {
 				allResources := []kube.Object{}
@@ -51,14 +38,14 @@ func GetResourcesForProfileName(profileName common.ProfileName, tier common.Odig
 	return nil, nil
 }
 
-func GetAvailableProfilesForTier(odigosTier common.OdigosTier) []commonprofiles.Profile {
+func GetAvailableProfilesForTier(odigosTier common.OdigosTier) []profile.Profile {
 	switch odigosTier {
 	case common.CommunityOdigosTier:
-		return GetAvailableCommunityProfiles()
+		return profiles.CommunityProfiles
 	case common.OnPremOdigosTier:
-		return GetAvailableOnPremProfiles()
+		return profiles.OnPremProfiles
 	default:
-		return []commonprofiles.Profile{}
+		return []profile.Profile{}
 	}
 }
