--- k8sutils/pkg/profiles/profile.go
@@ -1,165 +1,0 @@
-package profiles
-
-import (
-	odigosv1alpha1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
-	"github.com/odigos-io/odigos/common"
-	"github.com/odigos-io/odigos/profiles"
-)
-
-var (
-	// sizing profiles for the collectors resource settings
-	SizeSProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("size_s"),
-		ShortDescription: "Small size deployment profile",
-		ModifyConfigFunc: func(c *common.OdigosConfiguration) {
-			modifySizingConfig(c,
-				common.CollectorGatewayConfiguration{
-					MinReplicas:      1,
-					MaxReplicas:      5,
-					RequestCPUm:      150,
-					LimitCPUm:        300,
-					RequestMemoryMiB: 300,
-					LimitMemoryMiB:   300,
-				},
-				common.CollectorNodeConfiguration{
-					RequestMemoryMiB: 150,
-					LimitMemoryMiB:   300,
-					RequestCPUm:      150,
-					LimitCPUm:        300,
-				})
-		},
-	}
-	SizeMProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("size_m"),
-		ShortDescription: "Medium size deployment profile",
-		ModifyConfigFunc: func(c *common.OdigosConfiguration) {
-			modifySizingConfig(c,
-				common.CollectorGatewayConfiguration{
-					MinReplicas:      2,
-					MaxReplicas:      8,
-					RequestCPUm:      500,
-					LimitCPUm:        1000,
-					RequestMemoryMiB: 500,
-					LimitMemoryMiB:   600,
-				},
-				common.CollectorNodeConfiguration{
-					RequestMemoryMiB: 250,
-					LimitMemoryMiB:   500,
-					RequestCPUm:      250,
-					LimitCPUm:        500,
-				})
-		},
-	}
-	SizeLProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("size_l"),
-		ShortDescription: "Large size deployment profile",
-		ModifyConfigFunc: func(c *common.OdigosConfiguration) {
-			modifySizingConfig(c,
-				common.CollectorGatewayConfiguration{
-					MinReplicas:      3,
-					MaxReplicas:      12,
-					RequestCPUm:      750,
-					LimitCPUm:        1250,
-					RequestMemoryMiB: 750,
-					LimitMemoryMiB:   850,
-				},
-				common.CollectorNodeConfiguration{
-					RequestMemoryMiB: 500,
-					LimitMemoryMiB:   750,
-					RequestCPUm:      500,
-					LimitCPUm:        750,
-				})
-		},
-	}
-	AllowConcurrentAgents = profiles.Profile{
-		ProfileName:      common.ProfileName("allow_concurrent_agents"),
-		ShortDescription: "This profile allows Odigos to run concurrently with other agents",
-		ModifyConfigFunc: func(c *common.OdigosConfiguration) {
-			if c.AllowConcurrentAgents == nil {
-				allowConcurrentAgents := true
-				c.AllowConcurrentAgents = &allowConcurrentAgents
-			}
-		},
-	}
-	FullPayloadCollectionProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("full-payload-collection"),
-		ShortDescription: "Collect any payload from the cluster where supported with default settings",
-		KubeObject:       &odigosv1alpha1.InstrumentationRule{},
-	}
-	DbPayloadCollectionProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("db-payload-collection"),
-		ShortDescription: "Collect db payload from the cluster where supported with default settings",
-		KubeObject:       &odigosv1alpha1.InstrumentationRule{},
-	}
-	QueryOperationDetector = profiles.Profile{
-		ProfileName:      common.ProfileName("query-operation-detector"),
-		ShortDescription: "Detect the SQL operation name from the query text",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	SemconvUpgraderProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("semconv"),
-		ShortDescription: "Upgrade and align some attribute names to a newer version of the OpenTelemetry semantic conventions",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	CategoryAttributesProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("category-attributes"),
-		ShortDescription: "Add category attributes to the spans",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	CopyScopeProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("copy-scope"),
-		ShortDescription: "Copy the scope name into a separate attribute for backends that do not support scopes",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	HostnameAsPodNameProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("hostname-as-podname"),
-		ShortDescription: "Populate the spans resource `host.name` attribute with value of `k8s.pod.name`",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	JavaNativeInstrumentationsProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("java-native-instrumentations"),
-		ShortDescription: "Deprecated, native instrumentations are now enabled by default",
-	}
-	JavaEbpfInstrumentationsProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("java-ebpf-instrumentations"),
-		ShortDescription: "Instrument Java applications using eBPF instrumentation and eBPF enterprise processing",
-		KubeObject:       &odigosv1alpha1.InstrumentationRule{},
-	}
-	CodeAttributesProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("code-attributes"),
-		ShortDescription: "Record span attributes in 'code' namespace where supported",
-	}
-	DisableNameProcessorProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("disable-name-processor"),
-		ShortDescription: "If not using dotnet or java native instrumentations, disable the name processor which is not needed",
-	}
-	SmallBatchesProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("small-batches"),
-		ShortDescription: "Reduce the batch size for exports",
-		KubeObject:       &odigosv1alpha1.Processor{},
-	}
-	KratosProfile = profiles.Profile{
-		ProfileName:      common.ProfileName("kratos"),
-		ShortDescription: "Bundle profile that includes db-payload-collection, semconv, category-attributes, copy-scope, hostname-as-podname, code-attributes, query-operation-detector, disableNameProcessorProfile, small-batches, size_m, allow_concurrent_agents",
-		Dependencies:     []common.ProfileName{"db-payload-collection", "semconv", "category-attributes", "copy-scope", "hostname-as-podname", "code-attributes", "query-operation-detector", "disableNameProcessorProfile", "small-batches", "size_m", "allow_concurrent_agents"},
-	}
-	ProfilesMap = map[common.ProfileName]profiles.Profile{
-		SizeSProfile.ProfileName:                      SizeSProfile,
-		SizeMProfile.ProfileName:                      SizeMProfile,
-		SizeLProfile.ProfileName:                      SizeLProfile,
-		FullPayloadCollectionProfile.ProfileName:      FullPayloadCollectionProfile,
-		DbPayloadCollectionProfile.ProfileName:        DbPayloadCollectionProfile,
-		QueryOperationDetector.ProfileName:            QueryOperationDetector,
-		SemconvUpgraderProfile.ProfileName:            SemconvUpgraderProfile,
-		CategoryAttributesProfile.ProfileName:         CategoryAttributesProfile,
-		CopyScopeProfile.ProfileName:                  CopyScopeProfile,
-		HostnameAsPodNameProfile.ProfileName:          HostnameAsPodNameProfile,
-		JavaNativeInstrumentationsProfile.ProfileName: JavaNativeInstrumentationsProfile,
-		JavaEbpfInstrumentationsProfile.ProfileName:   JavaEbpfInstrumentationsProfile,
-		CodeAttributesProfile.ProfileName:             CodeAttributesProfile,
-		DisableNameProcessorProfile.ProfileName:       DisableNameProcessorProfile,
-		SmallBatchesProfile.ProfileName:               SmallBatchesProfile,
-		KratosProfile.ProfileName:                     KratosProfile,
-		AllowConcurrentAgents.ProfileName:             AllowConcurrentAgents,
-	}
-)
