+++ profiles/profile/profile.go
@@ -1,4 +1,4 @@
-package profiles
+package profile
 
 import (
 	"github.com/odigos-io/odigos/common"
@@ -9,6 +9,7 @@ import (
 
 type Profile struct {
 	ProfileName      common.ProfileName
+	MinimumTier      common.OdigosTier
 	ShortDescription string
 	KubeObject       K8sObject                         // used to read it from the embedded YAML file
 	Dependencies     []common.ProfileName              // other profiles that are applied by the current profile
